pyproject.toml
[project]
name = "adk-teams"
version = "0.1.0"
description = "Multi-agent critique fleet framework using Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.28.0",
    "litellm>=1.82.6",
    "pytest>=9.0.2",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.pytest_dotenv]
env_file = ".env"
